python-server/requirements.txt
connexion == 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
python-dotenv >= 0.19.0